Move get_wave_for_each item to waves.py (feels like it fits better there)
